Plumbed `pg_server_prepare` into DB::Pg via Rose::DB->register_db
This commit claims that `pg_server_prepare` is a boolean value with
no default. However, the documentation for DBD::Pg->prepare shows
that this is actually a trinary option:
https://metacpan.org/pod/DBD::Pg#prepare

  0: don't use prepared statements
  1: use prepared statements for SELECT, INSERT, UPDATE, and DELETE
  2: magic value for Postgresql 7.4 saying "prepare if you can"

There is no default value for this setting. This allows DBD::Pg to
choose its own default value. DBD::Pg should choose the `1` value
(unless running against Postgresql 7.4, in which case the `2` value
should be chosen).

Because of the mismatch between a `boolean` data type and the trinary
values available, this patch disallows explicitly sending the value
`2`. Changing from a `boolean` to a `scalar` data type would fix that,
though simply not specifying `pg_server_prepare` during `->register_db`
when using Rose::DB on Postgresql 7.4 should also accomplish sending `2`.